requirements.txt
Flask==1.0.2
Frozen-Flask==0.15
Jinja2==2.10
PyYAML==4.1
